restructure program
