+++ 2-Novikov-D21/HamiltonianPath.h
@@ -1,0 +1,2 @@
+#pragma once
+void FindHamiltonianPath();
